requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mutagen==1.47.0
python-dotenv==1.0.0
Pillow==10.0.0
